+++ Code/GameObject.java
@@ -5,12 +5,12 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-    protected int x, y;
+    protected float x, y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
     //Konstruktor für GameObject (wird als super Aufruf von anderen Klassen benutzt)
-    public GameObject (int x, int y, ID id) {
+    public GameObject (float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -40,11 +40,11 @@ public abstract class GameObject {
         this.velY = velY;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
@@ -52,11 +52,11 @@ public abstract class GameObject {
         return id;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
 }
